[DRIVER] - Rotary Encoder HY040 driver --- - Embedded Hal Based Driver for a HY040 encoder - Provide Direction   - Clockwise   - CounterClockwise   - Rest - Unit test to test the logic using Mocked pins
This works best when polling via a time based interrupt every 5ms.
However, the precision seem to be not the greatest (when reading the atomic, the value is often increased by 2 instead of one).
Tried a variant with interrupt which led to the same result.
Different things can be tested once modules that are supposed to
interpret the rotation are defined and implemented, but this allows to
move forward for now.

related to #13